color code logs
