+++ freemocap/data_layer/data_saver/data_models.py
@@ -11,11 +11,17 @@ from freemocap.utilities.create_nested_dict_from_pydantic import create_nested_d
 
 
 class Point(BaseModel):
+    """
+    A point in 3D space at a particular time
+    """
     x: Optional[float] = Field(None, description="The X-coordinate of the point")
     y: Optional[float] = Field(None, description="The Y-coordinate of the point")
     z: Optional[float] = Field(None, description="The Z-coordinate of the point")
 
 class VirtualMarkerDefinition(BaseModel):
+    """
+    A virtual marker, defined by combining multiple markers with weights to generate a new marker/point
+    """
     marker_names: List[str] = Field(default_factory=list, description="The names of the markers that define this virtual marker")
     marker_weights: List[float] = Field(default_factory=list, description="The weights of the markers that define this virtual marker, must sum to 1")
 
@@ -26,24 +32,27 @@ class VirtualMarkerDefinition(BaseModel):
             raise ValueError(f"Marker weights must sum to 1, got {marker_weights}")
         return values
 
-class Schema(BaseModel):
+class SegmentSchema(BaseModel):
+    """
+    A schema for a segment of a skeleton, defined by a set of tracked points and connections between them
+    """
     point_names: List[str] = Field(default_factory=list, description="The names of the tracked points that define this segment")
     virtual_marker_definitions: Optional[Dict[str, VirtualMarkerDefinition]] = Field(default_factory=dict, description="The virtual markers that define this segment")
     connections: List[Tuple[int, int]] = Field(default_factory=list, description="The connections between the tracked points")
     parent: Optional[str] = Field(None, description="The name of the parent of this segment")
 
 class SkeletonSchema(BaseModel):
-    body:  Schema = Field(default_factory=Schema, description="The tracked points that define the body")
-    hands: Dict[str, Schema] = Field(default_factory=dict, description="The tracked points that define the hands: keys - (left, right)")
-    face: Schema = Field(default_factory=Schema, description="The tracked points that define the face")
+    body:  SegmentSchema = Field(default_factory=SegmentSchema, description="The tracked points that define the body")
+    hands: Dict[str, SegmentSchema] = Field(default_factory=dict, description="The tracked points that define the hands: keys - (left, right)")
+    face: SegmentSchema = Field(default_factory=SegmentSchema, description="The tracked points that define the face")
 
     def __init__(self, schema_dict: Dict[str, Dict[str, Any]]):
         super().__init__()
-        self.body = Schema(**schema_dict["body"])
-        self.hands = {hand: Schema(**hand_schema) for hand, hand_schema in schema_dict["hands"].items()}
-        self.face = Schema(**schema_dict["face"])
+        self.body = SegmentSchema(**schema_dict["body"])
+        self.hands = {hand: SegmentSchema(**hand_schema) for hand, hand_schema in schema_dict["hands"].items()}
+        self.face = SegmentSchema(**schema_dict["face"])
 
-    def to_dict(self):
+    def dict(self):
         d = {}
         d["body"] = self.body.dict()
         d["hands"] = {hand: hand_schema.dict() for hand, hand_schema in self.hands.items()}
@@ -57,6 +66,9 @@ class Timestamps(BaseModel):
     by_camera: Dict[str, Any] = Field(default_factory=dict, description="Timestamps for each camera on this frame (key is video name)")
 
 class FrameData(BaseModel):
+    """
+    The data for a single frame
+    """
     timestamps: Timestamps = Field(default_factory=Timestamps, description="Timestamp data")
     tracked_points: Dict[str, Point] = Field(default_factory=dict, description="The points being tracked")
     @property
@@ -74,6 +86,9 @@ class FrameData(BaseModel):
         return d
 
 class InfoDict(BaseModel):
+    """
+    A dictionary of information about this recording, such as the measured segement lengths and the schemas that we can use to interpret the tracked points (i.e./e.g. how to connect the dots of skeleton)
+    """
     segment_lengths: Dict[str, Any] = Field(default_factory=dict, description="The lengths of the segments of the body")
     schemas: List[BaseModel] = Field(default_factory=list, description="The schemas for the tracked points")
 
@@ -86,5 +101,5 @@ if __name__ == "__main__":
 
     skeleton_schema = SkeletonSchema(schema_dict=mediapipe_skeleton_schema)
 
-    pprint.pp(skeleton_schema.to_dict())
+    pprint.pp(skeleton_schema.dict())
